new features

+++ src/main/java/org/example/dto/external/NBRBExchangeRateDto.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class NBRBExchangeRateDto {
@@ -15,10 +17,10 @@ public class NBRBExchangeRateDto {
     @JsonProperty("Cur_Abbreviation")
     private String fromCurrency;
     @JsonProperty("Cur_Scale")
-    private double scale;
+    private BigDecimal scale;
     @JsonProperty("Cur_Name")
     private String currencyName;
     @JsonProperty("Cur_OfficialRate")
-    private double exchangeRate;
+    private BigDecimal exchangeRate;
 
 }
